rewires dashboard to use separate routes.
