+++ addon/common-routes.js
@@ -1,5 +1,17 @@
 export function dashboardRoutes(router) {
-  router.route('dashboard');
+  router.route(
+    'dashboard',
+    {
+      path: 'dashboard',
+      resetNamespace: true,
+    },
+    function () {
+      this.route('index');
+      this.route('activities');
+      this.route('materials');
+      this.route('calendar');
+    }
+  );
   router.route('events', { path: 'events/:slug' });
   router.route('weeklyevents');
 }
