+++ addon/controllers/dashboard/activities.js
@@ -1,0 +1,3 @@
+import Controller from '@ember/controller';
+
+export default class DashboardActivitiesController extends Controller {}
