+++ addon/controllers/dashboard/index.js
@@ -1,0 +1,3 @@
+import Controller from '@ember/controller';
+
+export default class DashboardIndexController extends Controller {}
