+++ addon/controllers/dashboard/materials.js
@@ -1,0 +1,3 @@
+import Controller from '@ember/controller';
+
+export default class DashboardMaterialsController extends Controller {}
