+++ addon/routes/dashboard.js
@@ -1,0 +1,7 @@
+import Route from '@ember/routing/route';
+
+export default class DashboardRoute extends Route {
+  beforeModel() {
+    this.replaceWith('dashboard.index');
+  }
+}
