+++ addon/routes/dashboard/calendar.js
@@ -1,0 +1,3 @@
+import Route from '@ember/routing/route';
+
+export default class DashboardCalendarRoute extends Route {}
