+++ addon/routes/dashboard/index.js
@@ -1,0 +1,3 @@
+import Route from '@ember/routing/route';
+
+export default class DashboardIndexRoute extends Route {}
