+++ addon/routes/dashboard/materials.js
@@ -1,0 +1,3 @@
+import Route from '@ember/routing/route';
+
+export default class DashboardMaterialsRoute extends Route {}
