+++ addon/templates/dashboard/activities.hbs
@@ -1,0 +1,1 @@
+<DashboardAgenda />
