+++ addon/templates/dashboard/calendar.hbs
@@ -1,0 +1,32 @@
+<DashboardCalendar
+  @allSchools={{this.model.schools}}
+  @selectedDate={{this.selectedDate}}
+  @selectedView={{this.selectedView}}
+  @mySchedule={{this.mySchedule}}
+  @showFilters={{this.showFilters}}
+  @changeDate={{this.changeDate}}
+  @changeView={{set this.view}}
+  @selectEvent={{this.selectEvent}}
+  @toggleMySchedule={{this.toggleMySchedule}}
+  @toggleShowFilters={{this.toggleShowFilters}}
+  @selectedSchool={{this.selectedSchool}}
+  @changeSchool={{set this.school}}
+  @courseFilters={{this.courseFilters}}
+  @toggleCourseFilters={{fn (set this.courseFilters) (not this.courseFilters)}}
+  @selectedCohortIds={{split "-" this.cohorts}}
+  @addCohortId={{fn this.add "cohorts"}}
+  @removeCohortId={{fn this.remove "cohorts"}}
+  @selectedCourseLevels={{split "-" this.courseLevels}}
+  @addCourseLevel={{fn this.add "courseLevels"}}
+  @removeCourseLevel={{fn this.remove "courseLevels"}}
+  @selectedCourseIds={{split "-" this.courses}}
+  @addCourseId={{fn this.add "courses"}}
+  @removeCourseId={{fn this.remove "courses"}}
+  @selectedSessionTypeIds={{split "-" this.sessionTypes}}
+  @addSessionTypeId={{fn this.add "sessionTypes"}}
+  @removeSessionTypeId={{fn this.remove "sessionTypes"}}
+  @selectedTermIds={{split "-" this.terms}}
+  @addTermId={{fn this.add "terms"}}
+  @removeTermId={{fn this.remove "terms"}}
+  @clearFilters={{this.clearFilters}}
+/>
