+++ addon/templates/dashboard/index.hbs
@@ -1,0 +1,1 @@
+<DashboardWeek />
