+++ addon/templates/dashboard/materials.hbs
@@ -1,0 +1,1 @@
+<DashboardMaterials />
