+++ app/controllers/dashboard/activities.js
@@ -1,0 +1,1 @@
+export { default } from 'ilios-common/controllers/dashboard/activities';
