+++ app/routes/dashboard/index.js
@@ -1,0 +1,1 @@
+export { default } from 'ilios-common/routes/dashboard/index';
