+++ app/styles/ilios-common/components/dashboard-view-picker.scss
@@ -20,7 +20,7 @@
     }
   }
 
-  button {
+  a {
     @include ilios-button;
     padding: .25rem .5rem;
