+++ app/templates/dashboard/materials.js
@@ -1,0 +1,1 @@
+export { default } from 'ilios-common/templates/dashboard/materials';
