+++ tests/unit/routes/dashboard/calendar-test.js
@@ -1,0 +1,11 @@
+import { module, test } from 'qunit';
+import { setupTest } from 'ember-qunit';
+
+module('Unit | Route | dashboard/calendar', function (hooks) {
+  setupTest(hooks);
+
+  test('it exists', function (assert) {
+    let route = this.owner.lookup('route:dashboard/calendar');
+    assert.ok(route);
+  });
+});
