+++ tests/unit/routes/dashboard/materials-test.js
@@ -1,0 +1,11 @@
+import { module, test } from 'qunit';
+import { setupTest } from 'ember-qunit';
+
+module('Unit | Route | dashboard/materials', function (hooks) {
+  setupTest(hooks);
+
+  test('it exists', function (assert) {
+    let route = this.owner.lookup('route:dashboard/materials');
+    assert.ok(route);
+  });
+});
